Rename `interpret_callbacks.py` to `interpret.py`
